+++ packages/core/src/types/response.ts
@@ -5,28 +5,24 @@ import { InternalRoute } from "./route";
 export type RawParams = { [key: string]: string };
 export type Params = { [key: string]: any };
 
-export interface Response {
+export interface GenericResponse {
   location: HickoryLocation;
   key: string;
-  params: Params;
-  name: string;
-  partials: Array<string>;
   status: number;
+  title: string;
+  name: string;
+  params: Params;
+  partials: Array<string>;
   body: any;
   data: any;
-  title: string;
   error?: any;
   redirectTo?: ToArgument;
 }
+
+export type Response = GenericResponse;
 
 export interface Resolved {
   error: any;
   initial: any;
   every: any;
-}
-
-export interface PendingResponse {
-  resolved: Resolved;
-  route: InternalRoute;
-  response: Response;
 }