+++ packages/core/src/types/route.ts
@@ -31,7 +31,7 @@ export interface ResponseSetters {
 }
 
 export interface ResponseBuilder extends MatchedRouteProps {
-  resolved: Resolved;
+  resolved: Resolved | null;
   set: ResponseSetters;
   route: Interactions;
 }
@@ -65,10 +65,10 @@ export interface Route {
   path: string;
   keys: Array<string | number>;
   on: OnFns;
-  extra: { [key: string]: any };
+  extra?: { [key: string]: any };
 }
 
-export interface InternalMatch {
+export interface PathMatching {
   mustBeExact: boolean;
   re: RegExp;
   keys: Array<Key>;
@@ -77,17 +77,7 @@ export interface InternalMatch {
 export interface InternalRoute {
   public: Route;
   children: Array<InternalRoute>;
-  response: ResponseFn;
-  pathMatching: InternalMatch;
-  paramParsers: ParamParsers;
-}
-
-export interface MatchingRoute {
-  route: InternalRoute;
-  params: Params;
-}
-
-export interface BestMatch {
-  route: InternalRoute;
-  response: Response;
+  response?: ResponseFn;
+  pathMatching: PathMatching;
+  paramParsers?: ParamParsers;
 }