+++ packages/core/src/utils/once.ts
@@ -1,5 +1,5 @@
 export default function once(fn: () => Promise<any>): () => Promise<any> {
-  let promise: Promise<any> = null;
+  let promise: Promise<any>;
   let hasRun: boolean = false;
 
   return function(): Promise<any> {