+++ packages/core/types/resolveMatchedRoute.d.ts
@@ -1,0 +1,3 @@
+import { Match } from "./types/match";
+import { Resolved } from "./types/response";
+export default function resolveRoute(match: Match): Promise<Resolved>;