+++ packages/core/types/types/response.d.ts
@@ -1,31 +1,26 @@
 import { HickoryLocation, ToArgument } from "@hickory/root";
-import { InternalRoute } from "./route";
 export declare type RawParams = {
     [key: string]: string;
 };
 export declare type Params = {
     [key: string]: any;
 };
-export interface Response {
+export interface GenericResponse {
     location: HickoryLocation;
     key: string;
-    params: Params;
-    name: string;
-    partials: Array<string>;
     status: number;
+    title: string;
+    name: string;
+    params: Params;
+    partials: Array<string>;
     body: any;
     data: any;
-    title: string;
     error?: any;
     redirectTo?: ToArgument;
 }
+export declare type Response = GenericResponse;
 export interface Resolved {
     error: any;
     initial: any;
     every: any;
-}
-export interface PendingResponse {
-    resolved: Resolved;
-    route: InternalRoute;
-    response: Response;
 }