+++ packages/core/types/utils/matchProperties.d.ts
@@ -1,3 +1,0 @@
-import { MatchedRouteProps } from "../types/route";
-import { Response } from "../types/response";
-export default function matchProperties(response: Response): MatchedRouteProps;