+++ packages/vue/tests/Block.spec.ts
@@ -12,7 +12,10 @@ describe("Block component", () => {
   history.confirmWith = mockConfirmWith;
   history.removeConfirmation = mockRemoveConfirmation;
 
-  const routes = [{ name: "Place", path: "/place/:name" }];
+  const routes = [
+    { name: "Place", path: "/place/:name" },
+    { name: "Catch All", path: "(.*)" }
+  ];
   const router = curi(history, routes);
 
   const Vue = createLocalVue();